+++ src/pages/Club/ClubListPage/components/ClubSearchBarModal/ClubSearchBarModal.module.scss
@@ -30,53 +30,4 @@
       z-index: 1001;
     }
   }
-
-  &__input {
-    box-sizing: border-box;
-    border: solid 1px #4590bb;
-    border-radius: 4px;
-    height: 45px;
-    width: 100%;
-    font-size: 15px;
-    line-height: 160%;
-    padding: 12px 16px;
-    background-color: #f5f5f5;
-
-    @include media.media-breakpoint(mobile) {
-      padding-left: 10px;
-      margin-top: 18px;
-      border: none;
-      border-bottom: 1px solid #e1e1e1;
-      width: 100%;
-      height: 40px;
-      font-size: 14px;
-      outline: none;
-    }
-
-    &::placeholder {
-      color: #727272;
-
-      @include media.media-breakpoint(mobile) {
-        color: #727272;
-      }
-    }
-  }
-}
-
-.input-container {
-  background-color: #f5f5f5;
-
-  @include media.media-breakpoint(mobile) {
-    width: calc(100% - 40px);
-    border-radius: 16px;
-    padding: 1px 4%;
-    background-color: #f5f5f5;
-    display: flex;
-  }
-}
-
-.result {
-  @include media.media-breakpoint(mobile) {
-    padding: 0 5%;
-  }
 }
