+++ src/pages/Club/ClubListPage/components/ClubSearchContainer/ClubSearchContainer.module.scss
@@ -1,0 +1,38 @@
+@use "src/utils/scss/media";
+
+.search-bar {
+  box-sizing: border-box;
+  border: solid 1px #4590bb;
+  border-radius: 4px;
+  height: 45px;
+  width: 100%;
+  font-size: 15px;
+  line-height: 160%;
+  padding: 12px 16px;
+  background-color: #f5f5f5;
+
+  @include media.media-breakpoint(mobile) {
+    padding-left: 10px;
+    margin-top: 18px;
+    border: none;
+    border-bottom: 1px solid #e1e1e1;
+    width: 100%;
+    height: 40px;
+    font-size: 14px;
+    outline: none;
+  }
+
+  &::placeholder {
+    color: #727272;
+
+    @include media.media-breakpoint(mobile) {
+      color: #727272;
+    }
+  }
+}
+
+.result {
+  @include media.media-breakpoint(mobile) {
+    padding: 0 5%;
+  }
+}
